requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.3
colorama==0.4.6
fastapi==0.136.1
h11==0.16.0
idna==3.13
Jinja2==3.1.6
MarkupSafe==3.0.3
platformdirs==4.9.6
pydantic==2.13.4
pydantic_core==2.46.4
pyoxigraph==0.5.8
pyparsing==3.3.2
python-dotenv==1.2.2
python-multipart==0.0.27
rdflib==7.6.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
